requirements.txt
# requirements.txt
# Core ML and CV Libraries
tensorflow>=2.10.0
numpy>=1.22.0
opencv-python-headless>=4.5.5
Pillow>=9.0.0

# Web Framework
streamlit>=1.12.0

# Logging and Utilities
logging>=0.5.1.2

# Optional: For advanced tracking and analytics
scipy>=1.8.0
scikit-learn>=1.1.0

# Development and Testing
pytest>=7.0.0
mypy>=0.950